feat(repo): Use ruff
I cleaned up the configs and got rid of setup.cfg altogether.
To keep it simple, I use some very new features which will only be
available with the NEXT pip release in april.
I still consider this WIP. I want to test in a docker container and
extend CONTRIBUTING.md to explain both a pip and a uv workflow.